+++ src/Twilio/Rest/Conversations/V1/Configuration/WebhookOptions.php
@@ -12,9 +12,6 @@ namespace Twilio\Rest\Conversations\V1\Configuration;
 use Twilio\Options;
 use Twilio\Values;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- */
 abstract class WebhookOptions {
     /**
      * @param string $method The HTTP method to be used when sending a webhook