+++ src/Twilio/Rest/Conversations/V1/ConfigurationInstance.php
@@ -16,8 +16,6 @@ use Twilio\Values;
 use Twilio\Version;
 
 /**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- *
  * @property string $accountSid
  * @property string $defaultChatServiceSid
  * @property string $defaultMessagingServiceSid