+++ src/Twilio/Rest/Conversations/V1/Conversation/ParticipantList.php
@@ -17,9 +17,6 @@ use Twilio\Stream;
 use Twilio\Values;
 use Twilio\Version;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- */
 class ParticipantList extends ListResource {
     /**
      * Construct the ParticipantList