+++ src/Twilio/Rest/Conversations/V1/Service/Configuration/NotificationList.php
@@ -12,9 +12,6 @@ namespace Twilio\Rest\Conversations\V1\Service\Configuration;
 use Twilio\ListResource;
 use Twilio\Version;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- */
 class NotificationList extends ListResource {
     /**
      * Construct the NotificationList