+++ src/Twilio/Rest/Conversations/V1/Service/Conversation/MessageContext.php
@@ -19,8 +19,6 @@ use Twilio\Values;
 use Twilio\Version;
 
 /**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- *
  * @property DeliveryReceiptList $deliveryReceipts
  * @method \Twilio\Rest\Conversations\V1\Service\Conversation\Message\DeliveryReceiptContext deliveryReceipts(string $sid)
  */