+++ src/Twilio/Rest/Conversations/V1/Service/ConversationContext.php
@@ -21,8 +21,6 @@ use Twilio\Values;
 use Twilio\Version;
 
 /**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- *
  * @property ParticipantList $participants
  * @property MessageList $messages
  * @property WebhookList $webhooks