+++ src/Twilio/Rest/Conversations/V1/Service/ConversationOptions.php
@@ -12,9 +12,6 @@ namespace Twilio\Rest\Conversations\V1\Service;
 use Twilio\Options;
 use Twilio\Values;
 
-/**
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
- */
 abstract class ConversationOptions {
     /**
      * @param string $friendlyName The human-readable name of this conversation.